更改下载图片不正确的问题

+++ HealthExpo/LiuYanModul.m
@@ -12,7 +12,7 @@
 @implementation LiuYanModul
 
 - (void)getLiyanList:(NSInteger)cid{
-    NSString *urlPath = [NSString stringWithFormat:@"/mobile/getCommentInfoList.action?cid=%ld",cid];
+    NSString *urlPath = [NSString stringWithFormat:@"/mobile/getCommentInfoList.action?cid=%ld",(long)cid];
     HENetTask *task = [[HENetTask alloc] initWithUrlString:urlPath];
     __weak __typeof(self) weakSelf = self;
     task.successBlock = ^(NSURLSessionDataTask *task, NSArray *responseObject) {
